main added its own v11 migration (imported_pool_block_count and
imported_epoch_block_total). The runner keys migration state by version and
validates the recorded name and checksum per version, so two migrations
claiming v11 is rejected at startup with a checksum-drift error rather than
applied. This branch's ALTER TABLE pool_registration ADD COLUMN deposit_held
moves to v12: v11/sqlite/expand.sql keeps main's content, the branch's file
becomes v12/sqlite/expand.sql byte for byte, and the registry entry,
registry_test.go expectations, pool_deposit_held_test.go schema_migrations
version references and the DATABASE.md migration references move with it. Only
the sqlite directory ships SQL; postgres and mysql are translated from it. The
migration has no contract.sql.

applyPoolRegistrationCertificate takes deposit as *uint64 on main (#3836), so
poolRegistrationDepositHeld now takes and returns *uint64 and the column binds
through nullableDecimalUint64. An uncomputable deposit is stored NULL instead
of an authoritative zero, and reads back through the deposit_amount fallback a
row predating the column already takes.

+++ ledger/verify_opcert.go
@@ -60,29 +60,29 @@ func opCertFromHeader(header ledger.BlockHeader) (*ledger.OpCert, bool) {
 	case *shelley.ShelleyBlockHeader:
 		return shelleyOpCert(
 			h.Body.OpCertHotVkey,
-			h.Body.OpCertSequenceNumber,
-			h.Body.OpCertKesPeriod,
+			uint64(h.Body.OpCertSequenceNumber),
+			uint64(h.Body.OpCertKesPeriod),
 			h.Body.OpCertSignature,
 		), true
 	case *allegra.AllegraBlockHeader:
 		return shelleyOpCert(
 			h.Body.OpCertHotVkey,
-			h.Body.OpCertSequenceNumber,
-			h.Body.OpCertKesPeriod,
+			uint64(h.Body.OpCertSequenceNumber),
+			uint64(h.Body.OpCertKesPeriod),
 			h.Body.OpCertSignature,
 		), true
 	case *mary.MaryBlockHeader:
 		return shelleyOpCert(
 			h.Body.OpCertHotVkey,
-			h.Body.OpCertSequenceNumber,
-			h.Body.OpCertKesPeriod,
+			uint64(h.Body.OpCertSequenceNumber),
+			uint64(h.Body.OpCertKesPeriod),
 			h.Body.OpCertSignature,
 		), true
 	case *alonzo.AlonzoBlockHeader:
 		return shelleyOpCert(
 			h.Body.OpCertHotVkey,
-			h.Body.OpCertSequenceNumber,
-			h.Body.OpCertKesPeriod,
+			uint64(h.Body.OpCertSequenceNumber),
+			uint64(h.Body.OpCertKesPeriod),
 			h.Body.OpCertSignature,
 		), true
 	case *babbage.BabbageBlockHeader:
@@ -94,16 +94,22 @@ func opCertFromHeader(header ledger.BlockHeader) (*ledger.OpCert, bool) {
 	}
 }
 
+// shelleyOpCert takes the counter and KES period as uint64 because that is
+// what cardano-ledger decodes them as (Word64 and KESPeriod{Word}) and what
+// ledger.OpCert already carries. The TPraos header bodies they come from are
+// gouroboros types whose declared width is the release's to choose, so the
+// call sites convert; widening this signature keeps that conversion the only
+// place a release change is visible and keeps it lossless in either direction.
 func shelleyOpCert(
 	hotVkey []byte,
-	sequenceNumber uint32,
-	kesPeriod uint32,
+	sequenceNumber uint64,
+	kesPeriod uint64,
 	signature []byte,
 ) *ledger.OpCert {
 	return &ledger.OpCert{
 		KesVkey:       hotVkey,
-		IssueNumber:   uint64(sequenceNumber),
-		KesPeriod:     uint64(kesPeriod),
+		IssueNumber:   sequenceNumber,
+		KesPeriod:     kesPeriod,
 		ColdSignature: signature,
 	}
 }
